+++ contracts/interfaces/IRegistries.sol
@@ -4,10 +4,10 @@ pragma solidity ^0.8.18;
 interface IRegistries {
     struct Registry {
         // Percentage of a tickets value that will be rewarded to
-        // delegated stakers expressed as a fraction of 10000.
+        // delegated stakers expressed as a fraction of 100000.
         // This value is currently locked to the default payout percentage
         // until epochs are implemented.
-        uint16 payoutPercentage;
+        uint32 payoutPercentage;
         // Public http/s endpoint to retrieve additional metadata
         // about the node.
         // The current metadata schema is as follows:
@@ -23,7 +23,7 @@ interface IRegistries {
 
     function register(string calldata publicEndpoint) external;
 
-    function setDefaultPayoutPercentage(uint16 _defaultPayoutPercentage) external;
+    function setDefaultPayoutPercentage(uint32 _defaultPayoutPercentage) external;
 
     function setSeekerAccount(
         address seekerAccount,
